// mod01/src/app/models/staff.model.ts
export interface Staff {
    ID:number;
    Name: string;
    Country: string;
    Salary: number;
    Email: string;
}

// export class Staff{
//     Id:number;
//     Name: string;
//     Country: string;
//     Salary: number;
//     Email: string;

//     constructor(_id:number,_name:string,_country:string,_salary:number,_email:string){
//         this.Id = _id;
//         this.Name = _name;
//         this.Country = _country;
//         this.Salary = _salary;
//         this.Email = _email;
//     }

//     //用法:new Staff(val)、const list: Staff = {}

// }